Fix experts page by splitting up experts table components
